Conflicts:
	VolleyManagement.Dal.MsSql/Infrastructure/NinjectDataAccessModule.cs
	VolleyManagement.Dal.MsSql/Mappers/DomainToDal.cs
	VolleyManagement.Dal.MsSql/Sql/CreateDatabase.sql
	VolleyManagement.Dal.MsSql/Sql/InsertValues.sql
	VolleyManagement.Dal.MsSql/VolleyManagement.Dal.MsSql.csproj
	VolleyManagement.Dal.MsSql/VolleyManagementContext.cs
	VolleyManagement.Dal/VolleyManagement.Dal.csproj
	VolleyManagement.Services/Infrastructure/NinjectServiceBindModule.cs
	VolleyManagement.Services/VolleyManagement.Services.csproj
	VolleyManagement.UI/Areas/Mvc/Views/Home/Index.cshtml
	VolleyManagement.UI/Areas/WebAPI/WebAPIAreaRegistration.cs
	VolleyManagement.UI/VolleyManagement.UI.csproj
	VolleyManagement.UnitTests/Mvc/Controllers/PlayersControllerTests.cs
	VolleyManagement.UnitTests/VolleyManagement.UnitTests.csproj

+++ VolleyManagement.Services/PlayerService.cs
@@ -1,14 +1,16 @@
 ﻿namespace VolleyManagement.Services
 {
     using System;
+    using System.ComponentModel.DataAnnotations;
     using System.Linq;
     using VolleyManagement.Contracts;
     using VolleyManagement.Contracts.Exceptions;
     using VolleyManagement.Dal.Contracts;
     using VolleyManagement.Dal.Exceptions;
     using VolleyManagement.Domain.Players;
-
+    using VolleyManagement.Domain.Teams;
     using DAL = VolleyManagement.Dal.Contracts;
+    using IsolationLevel = System.Data.IsolationLevel;
 
     /// <summary>
     /// Defines PlayerService
@@ -20,13 +22,17 @@
         /// </summary>
         private readonly IPlayerRepository _playerRepository;
 
+        private readonly ITeamRepository _teamRepository;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="PlayerService"/> class.
         /// </summary>
-        /// <param name="playerRepository">The user repository</param>
-        public PlayerService(IPlayerRepository playerRepository)
+        /// <param name="playerRepository">The player repository</param>
+        /// <param name="teamRepository">The team repository</param>
+        public PlayerService(IPlayerRepository playerRepository, ITeamRepository teamRepository)
         {
             _playerRepository = playerRepository;
+            _teamRepository = teamRepository;
         }
 
         /// <summary>
@@ -44,8 +50,20 @@
         /// <param name="playerToCreate">A Player to create.</param>
         public void Create(Player playerToCreate)
         {
-            _playerRepository.Add(playerToCreate);
-            _playerRepository.UnitOfWork.Commit();
+            using (IDbTransaction transaction = _playerRepository.UnitOfWork.BeginTransaction(IsolationLevel.ReadUncommitted))
+            {
+                if (playerToCreate.TeamId != null)
+                {
+                    if (_teamRepository.FindWhere(t => t.Id == playerToCreate.TeamId).SingleOrDefault() == null)
+                    {
+                        throw new MissingEntityException(ServiceResources.ExceptionMessages.TeamNotFound, playerToCreate.TeamId);
+                    }
+                }
+
+                _playerRepository.Add(playerToCreate);
+                _playerRepository.UnitOfWork.Commit();
+                transaction.Commit();
+            }
         }
 
         /// <summary>
@@ -58,11 +76,11 @@
             Player player;
             try
             {
-                player = _playerRepository.FindWhere(t => t.Id == id).Single();
+                player = _playerRepository.FindWhere(p => p.Id == id).Single();
             }
             catch (InvalidOperationException ex)
             {
-                throw new MissingEntityException("Player with specified Id can not be found", ex);
+                throw new MissingEntityException(ServiceResources.ExceptionMessages.PlayerNotFound, ex);
             }
 
             return player;
@@ -74,16 +92,35 @@
         /// <param name="playerToEdit">Player to edit.</param>
         public void Edit(Player playerToEdit)
         {
-            try
+            using (IDbTransaction transaction = _playerRepository.UnitOfWork.BeginTransaction(IsolationLevel.ReadUncommitted))
             {
-                _playerRepository.Update(playerToEdit);
-            }
-            catch (InvalidKeyValueException ex)
-            {
-                throw new MissingEntityException("Player with specified Id can not be found", ex);
-            }
+                // Check if player is captain of team and teamId is null or changed
+                Team ledTeam = GetPlayerLedTeam(playerToEdit.Id);
+                if (ledTeam != null &&
+                    (playerToEdit.TeamId == null || playerToEdit.TeamId != ledTeam.Id))
+                {
+                    var ex = new ValidationException(ServiceResources.ExceptionMessages.PlayerIsCaptainOfAnotherTeam);
+                    ex.Data[Domain.Constants.ExceptionManagement.ENTITY_ID_KEY] = ledTeam.Id;
+                    throw ex;
+                }
+                else if (playerToEdit.TeamId != null
+                         && _teamRepository.FindWhere(t => t.Id == playerToEdit.TeamId).SingleOrDefault() == null)
+                {
+                    throw new MissingEntityException(ServiceResources.ExceptionMessages.TeamNotFound, playerToEdit.TeamId);
+                }
 
-            _playerRepository.UnitOfWork.Commit();
+                try
+                {
+                    _playerRepository.Update(playerToEdit);
+                }
+                catch (InvalidKeyValueException ex)
+                {
+                    throw new MissingEntityException(ServiceResources.ExceptionMessages.PlayerNotFound, ex);
+                }
+
+                _playerRepository.UnitOfWork.Commit();
+                transaction.Commit();
+            }
         }
 
         /// <summary>
@@ -92,6 +129,14 @@
         /// <param name="id">The id of player to delete.</param>
         public void Delete(int id)
         {
+            Team playerTeam = GetPlayerLedTeam(id);
+            if (playerTeam != null)
+            {
+                var ex = new ValidationException(ServiceResources.ExceptionMessages.PlayerIsCaptainOfAnotherTeam);
+                ex.Data[Domain.Constants.ExceptionManagement.ENTITY_ID_KEY] = playerTeam.Id;
+                throw ex;
+            }
+
             try
             {
                 _playerRepository.Remove(id);
@@ -99,9 +144,35 @@
             }
             catch (InvalidKeyValueException ex)
             {
-                var serviceException = new MissingEntityException("Player with specified Id can not be found", ex);
-                throw serviceException;
+                throw new MissingEntityException(ServiceResources.ExceptionMessages.PlayerNotFound, id, ex);
             }
+        }
+
+        /// <summary>
+        /// Find team of specified player
+        /// </summary>
+        /// <param name="player">Player which team should be found</param>
+        /// <returns>Player's team</returns>
+        public Team GetPlayerTeam(Player player)
+        {
+            if (player.TeamId == null)
+            {
+                return null;
+            }
+
+            try
+            {
+                return _teamRepository.FindWhere(t => t.Id == player.TeamId).Single();
+            }
+            catch (InvalidOperationException ex)
+            {
+                throw new MissingEntityException(ServiceResources.ExceptionMessages.TeamNotFound, player.TeamId, ex);
+            }
+        }
+
+        private Team GetPlayerLedTeam(int playerId)
+        {
+            return _teamRepository.FindWhere(t => t.CaptainId == playerId).SingleOrDefault();
         }
     }
 }
